requirements.txt
openai
graphviz
faiss-cpu
numpy
requests
python-dotenv
pymupdf
scikit-learn
sentence-transformers
